закомментирован код для ошибок парсинга, редактирование музея, экспоната,выставки, лайки

+++ db/src/main/java/museum/mapper/MuseumMapper.java
@@ -1,6 +1,5 @@
 package museum.mapper;
 
-import museum.domen.AuthorModel;
 import museum.domen.MuseumModel;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
@@ -11,8 +10,7 @@ public interface MuseumMapper extends CrudRepository<MuseumModel, Long> {
   //MuseumModel findMuseumModelByLogin(String login);
 
   //MuseumModel findMuseumModelByLoginAndId(String login, int id);
-  MuseumModel findById(Long id);
-
+  MuseumModel findById(Integer id);
 
   @Query("Select   m  From  museums  m join users  u on  u.museum = m.id WHERE m.id = :id and u.login =:loginAdmin   ")
   MuseumModel findMuseumByIdAndMuseumAdmin(@Param("id") Integer id, @Param("loginAdmin") String loginAdmin);
